+++ package.json
@@ -12,8 +12,8 @@
 		"typescript": "^5"
 	},
 	"dependencies": {
-		"@drizzle-team/brocli": "^0.11.0",
 		"arktype": "^2.1.22",
+		"gunshi": "^0.26.3",
 		"ink": "^6.2.3",
 		"react": "^19.1.1",
 		"toml": "^3.0.0"
